producer/requirements.txt
kafka-python==2.0.2
numpy==2.1.2
scipy==1.14.1
sympy==1.13.3